lead-crawl-cblc/company-leads/upload_company_list/requirements.txt
Flask==1.1.1
botocore==1.15.36
boto3==1.12.41
mysql_connector_repackaged==0.3.1
